rename trip_mode_share_calculator2.py to trip_mode_share_calculator.py
